move genres subquery to user resolver
