+++ projects/backend/src/services/services.ts
@@ -47,7 +47,7 @@ export const initServices = (config: IConfig, database: IDatabaseClient): IServi
 	const shareService = new ShareService(database);
 	const userService = new UserService(database);
 	const songTypeService = new SongTypeService(database);
-	const genreService = new GenreService(database);
+	const genreService = new GenreService(database, shareService);
 	const artistService = new ArtistService(songService, shareService);
 	const artistExtractor = new ArtistExtractor();
 	const songMetaDataService = new SongMetaDataService([
